Rename property classes for consistency.
Renames property_view to property, property_value_view to property_value
and property_value to encoded_property_value. Class names without the
"view" are more consistent with other class names such as "layer" and
"feature" which are also views in a way but don't have that in their
name. It also makes the names of classes more often used in user code
shorter.