Change the way the feature builder works.
The ID has to be set now by calling set_id() on the feature builder
instead of putting it in the constructor. This way we can decide
whether we want to have an ID or not (spec says IDs are optional).
The call to set_id() must be the first call after the construction
of the feature builder, before the geometry and then the properties
are added.