Add paranoia check for too large counts in geometry commands.
Geometries contain commands which have a count telling us how many
points are to be expected after this command. This count is often used
in users of vtzero to reserve enought memory for all points. An
attacker could craft a tiny vector tile containing a huge count value
and this could lead to memory problems in the user program. So this
commit adds a check that the count is not too large compared to the
number of bytes in the geometry. You can still create a huge tile to
trigger this problem but it doesn't work with a tiny tile any more.