Use index for key/values tables in layer builder when they get large.
The key/values tables built into the layer builders are flat tables.
Linear search is used to find entries in them which will be slower and
slower the larger the table grows. This adds an index (based on
std::unordered_map) that is populated once a table reaches a certain
size and used from then on.