Roll back features added without an explicit commit().
This completely turns the previous behaviour on its head. Before the
change a feature builder going out of scope would automatically commit
the feature. Now the feature is rolled back. You have to call commit()
explicitly before the feature is destroyed to commit it.

This is the more logical behaviour in an environment where exceptions
can occur, because any exception while writing a feature will roll back
this feature instead of committing a (possibly incomplete) one.