Fix skip_structured_value() and add programm for fuzz testing it.
The skip_structured_value() function did not handle number-list
attributes at all and had a bug in the handling of map attributes. Both
are now fixed.

In addition there is a new program test/fuzz/ft_skip_structured_value
that can be used with AFL to fuzz test this function.