Pass area to ring_end if geometry handler accepts it
We are calculating the ring area during decoding to determine whether the ring is an outer ring or inner ring. In some downstream applications, it's useful to know the area of the ring in addition to winding, so that it can e.g. filter small rings. This changes accepts geometry handlers with an alternative signature `ring_end(int64_t)` and calls them instead of `ring_end(ring_type)`.

Care has been taken to disallow geometry handlers that have _both_ signatures (you'll get a `call to 'classify_ring_area' is ambiguous` error message). Existing applications should continue to work as is, only accepting the ring type.